*write output in VTU (i.e. VTK Unstructured grid) format at runtime*
Summary:

write visualization data for your simulation results in VTK/VTU format
 + more flexible than with current post-processor / filters
 + at runtime
 + in parallel
 + using binary-coded files

+++ +++ WORK IN PROGRESS +++ CALL FOR USERS AND CONTRIBUTORS +++ +++

Details:

*writing data to filesystem in VTU format*

 - new abstract base class VtkWriterBase and derived class VtuWriter

 - these classes extract the purely VTK/VTU related functionality from
   the already existing classes used in post processing (VtkWriter/
   VtuWriter/..., derived from PostWriterBase)

 - the existing classes are renamed to better express their actual
   responsibility: VtkWriter -> PostVtkWriter, VtuWriter -> PostVtuWriter, ...

 - the new classes handle the output file streams including the formal xml
   file structure (header, footer, keywords, ...) as defined by VTK/VTU

 - to allow for parallel output, a master file is written (by processor 0) which
   summarizes the processor-individual output files

 - one set of files, i.e. one file per processor ('.vtu') and one master
   file ('.pvtu'), is written for each time step

 - all master files for a series of time steps can be summarized by an extra
   '.pvd' collection file which is then conveniently loaded into Paraview

 - the geometry of an unstructured grid is defined by cells (cell type,
   point offset, ...) and points (point coordinates)

 - visualization data may consist of n components per point or per cell

 - numeric data in output files may either be written in binary or ascii
   format

Note: The new classes should be used in post processing at some point to
      avoid code duplication. Due to missing tests and entanglement of
      functionalities and abstraction levels, I had a hard time trying to
      and finally failed at replacing the existing code fragments by the
      new ones though. If you are currently using VTK post processing
      and could contribute to this work: any help is appreciated :)

*writing data in VTU format at runtime*

 - new class RuntimeVtuWriter which most importantly holds a pointer
   to a VtuWriter object

 - additionally, it stores
   - the data that defines the geometry of an unstructured grid
   - all the desired visualization data in point / cell data vectors

 - the main idea is to flexibly (re)set the geometry of the grid and
   also append the visualization data one after another (possibly at
   different places in your code)

 - finally, upon request, the stored data is written to filesystem
   all at once

*writing visualization data for a discretization in VTU format at runtime*

 - new class DiscretizationRuntimeVtuWriter which holds a pointer to a
   RuntimeVtuWriter and a DRT::Dicretization

 - the geometry of the discretization is created from DRT::Elements and
   DRT::Nodes based on the reference configuration

 - the already existing list that relates the DiscretizationType
   ('Shape') of an element to a VTK cell type is moved to a central
   more place to avoid duplication

 - in this 'standard' case, one element will be represented by one VTK
   cell and the number of points equals the number of nodes

 - 'Dof based' and 'node based' results are handled as point data,
   whereas 'element based' results are handled as cell data

 - the strategy in parallel computing is that each processor only writes
   the geometry and data of its row elements

 - For the reasons stated above, unfortunately, the implementation
   is not finished yet.

*writing visualization data for beam elements in VTU format at runtime*

 - beam elements are not covered by the funtionality of the
   DiscretizationRuntimeWriter described above because
   a) cubic Hermite polynomials as used for beam centerline interpolation are
      not supported by VTK
   b) beam specific data such as the triad field or cross-section
      resultants require interaction with the elements themselves to
      generate/interpolate the data from a given global displacement state

 - new class BeamDiscretizationRuntimeVtuWriter which holds a pointer to a
   RuntimeVtuWriter and a DRT::Dicretization containing beam elements

 - the geometry of one beam element is subdivided in n segments and
   represented by a VTK_POLYLINE (piecewise straight line) with n+1 points

 - instead of the initial/reference configuration, also current positions can be
   used as point coordinates for the geometry definition of the
   unstructured grid (geometry must be written in every time step in this case)

 - already implemented 'special' visualization output:
   - triads (three orthonormal base vectors)
   - material cross-section strains at Gauss points: axial strain, shear
     strains, twist, curvatures
   - material cross-section stresses at Gauss points: axial force, shear
     forces, torque, bending moments
   - radius of a circular cross-section

*BeamBase and Beam3r*

 - new data members and methods to store and access the material
   cross-section strains and stresses for visualization purposes

*structure model evaluator structure*

 - check parameters and call VTK output method if required

 - for now, create and initialize BeamDiscretizationRuntimeVtuWriter
   in every time step

*definition and handling of input parameters*

 - added required input parameters to list of valid parameters

 - new parameter container classes ParamsRuntimeVtkOutput and BeamRuntimeVtuOutputParams
   which store and handle the parameter values

*storage and initialization of parameter containers*

 - parameter containers are initialized only if required

 - a pointer to a ParamsRuntimeVtkOutput object is stored as a member
   of BaseDataIO container if required

 - a pointer to a BeamRuntimeVtuOutputParams object is stored as a
   member of ParamsRuntimeVtkOutput container if required

*Miscellaneous*
 - added new cpp files to CMakeList

 - add const qualifiers for method GetCurrentElementDis() in beam
   interaction utils file

 - changes to tests
   - added parameters for VTU output exemplarily to one input file

 - fix compiler warnings:
   - Beam3k: avoid comparison of signed and unsigned integer
   - ale_meshtying: comment out two unused class variables

Special thanks to Martin K. for idea, discussions and support to get started!

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@23005 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_beam3/beam3k.cpp
@@ -1234,7 +1234,7 @@ void DRT::ELEMENTS::Beam3k::GetGeneralizedInterpolationMatrixVariationsAtXi(
   const unsigned int numdof = ndim*vpernode*nnodecl + BEAM3K_COLLOCATION_POINTS;
 
   // safety check
-  if ( Ivar.M() != 2*ndim or Ivar.N() != numdof )
+  if ( (unsigned int) Ivar.M() != 2*ndim or (unsigned int) Ivar.N() != numdof )
     dserror("size mismatch! expected %dx%d matrix and got %dx%d",6 ,numdof ,Ivar.M() ,Ivar.N() );
 
   Ivar.Zero();
@@ -1397,7 +1397,7 @@ void DRT::ELEMENTS::Beam3k::GetStiffmatResultingFromGeneralizedInterpolationMatr
   const unsigned int numdof = ndim*vpernode*nnodecl + BEAM3K_COLLOCATION_POINTS;
 
   // safety check
-  if ( stiffmat.M() != numdof or stiffmat.N() != numdof )
+  if ( (unsigned int) stiffmat.M() != numdof or (unsigned int) stiffmat.N() != numdof )
     dserror("size mismatch! expected %dx%d matrix and got %dx%d",
         numdof, numdof, stiffmat.M(), stiffmat.N() );
 
@@ -1551,7 +1551,7 @@ void DRT::ELEMENTS::Beam3k::GetGeneralizedInterpolationMatrixIncrementsAtXi(
   const unsigned int numdof = ndim*vpernode*nnodecl + BEAM3K_COLLOCATION_POINTS;
 
   // safety check
-  if ( Iinc.M() != 2*ndim or Iinc.N() != numdof )
+  if ( (unsigned int) Iinc.M() != 2*ndim or (unsigned int) Iinc.N() != numdof )
     dserror("size mismatch! expected %dx%d matrix and got %dx%d", 6, numdof, Iinc.M(), Iinc.N() );
 
   Iinc.Zero();
